+++ packages/amplify-graphql-api-construct/src/internal/default-parameters.ts
@@ -20,5 +20,4 @@ export const defaultTranslationBehavior: TranslationBehavior = {
   enableTransformerCfnOutputs: false,
   allowDestructiveGraphqlSchemaUpdates: false,
   replaceTableUponGsiUpdate: false,
-  _allowGen1Patterns: true,
 };