+++ packages/amplify-graphql-predictions-transformer/src/graphql-predictions-transformer.ts
@@ -157,6 +157,13 @@ export class PredictionsTransformer extends TransformerPluginBase {
       return;
     }
 
+    // This validation can't occur in validate because the api has not been initialized until generateResolvers
+    if (!context.transformParameters.allowGen1Patterns) {
+      cdk.Annotations.of(context.api).addWarning(
+        `@${PredictionsDirective.name} is deprecated. This functionality will be removed in the next major release.`,
+      );
+    }
+
     const stack: cdk.Stack = context.stackManager.createStack(PREDICTIONS_DIRECTIVE_STACK);
     const env = context.synthParameters.amplifyEnvironmentName;
     const createdResources = new Map<string, any>();