+++ packages/amplify-graphql-relational-transformer/src/graphql-belongs-to-transformer.ts
@@ -17,6 +17,7 @@ import {
   TransformerTransformSchemaStepContextProvider,
 } from '@aws-amplify/graphql-transformer-interfaces';
 import { BelongsToDirective } from '@aws-amplify/graphql-directives';
+import { Annotations } from 'aws-cdk-lib';
 import {
   DirectiveNode,
   DocumentNode,
@@ -150,6 +151,22 @@ export class BelongsToTransformer extends TransformerPluginBase {
     const context = ctx as TransformerContextProvider;
 
     for (const config of this.directiveList) {
+      // This validation can't occur in validate because the api has not been initialized until generateResolvers
+      if (!ctx.transformParameters.allowGen1Patterns) {
+        const { field, object } = config;
+        const modelName = object.name.value;
+        const fieldName = field.name.value;
+        if (field.type.kind === Kind.NON_NULL_TYPE) {
+          Annotations.of(ctx.api).addWarning(
+            `@${BelongsToDirective.name} on required fields is deprecated. Modify ${modelName}.${fieldName} to be optional. This functionality will be removed in the next major release.`,
+          );
+        }
+        if (config.fields) {
+          Annotations.of(ctx.api).addWarning(
+            `fields argument on @${BelongsToDirective.name} is deprecated. Modify ${modelName}.${fieldName} to use references instead. This functionality will be removed in the next major release.`,
+          );
+        }
+      }
       const dbType = getStrategyDbTypeFromTypeNode(config.field.type, context);
       const dataSourceBasedTransformer = getBelongsToDirectiveTransformer(dbType, config);
       dataSourceBasedTransformer.generateResolvers(ctx, config);
@@ -159,21 +176,6 @@ export class BelongsToTransformer extends TransformerPluginBase {
 
 const validate = (config: BelongsToDirectiveConfiguration, ctx: TransformerContextProvider): void => {
   const { field, object } = config;
-  if (!ctx.transformParameters.allowGen1Patterns) {
-    const modelName = object.name.value;
-    const fieldName = field.name.value;
-    if (field.type.kind === Kind.NON_NULL_TYPE) {
-      throw new InvalidDirectiveError(
-        `@${BelongsToDirective.name} cannot be used on required fields. Modify ${modelName}.${fieldName} to be optional.`,
-      );
-    }
-    if (config.fields) {
-      throw new InvalidDirectiveError(
-        `fields argument on @${BelongsToDirective.name} is disallowed. Modify ${modelName}.${fieldName} to use references instead.`,
-      );
-    }
-  }
-
   if (config.overrideIndexName && !config.references) {
     throw new InvalidDirectiveError(
       `overrideIndexName cannot be used on @${BelongsToDirective.name} without references. Modify ${object.name.value}.${field.name.value} to use references or remove overrideIndexName.`,