+++ packages/amplify-graphql-transformer-interfaces/src/graphql-api-provider.ts
@@ -4,6 +4,7 @@ import { Grant, IGrantable, IRole } from 'aws-cdk-lib/aws-iam';
 // eslint-disable-next-line import/no-cycle
 import { TransformHostProvider } from './transform-host-provider';
 import { AssetProvider } from './asset-provider';
+import { IamResource } from 'aws-cdk-lib/aws-appsync';
 
 // Auth Config Modes
 export type AppSyncAuthMode = 'API_KEY' | 'AMAZON_COGNITO_USER_POOLS' | 'AWS_IAM' | 'OPENID_CONNECT' | 'AWS_LAMBDA';
@@ -127,6 +128,7 @@ export type MappingTemplateProvider = InlineMappingTemplateProvider | S3MappingT
 
 export interface GraphQLAPIProvider extends IConstruct {
   readonly apiId: string;
+  readonly graphqlUrl: string;
   readonly host: TransformHostProvider;
   readonly name: string;
   readonly assetProvider: AssetProvider;
@@ -136,7 +138,7 @@ export interface GraphQLAPIProvider extends IConstruct {
   addToSchema: (addition: string) => void;
   addSchemaDependency: (construct: CfnResource) => boolean;
 
-  grant: (grantee: IGrantable, resources: APIIAMResourceProvider, ...actions: string[]) => Grant;
+  grant: (grantee: IGrantable, resources: IamResource, ...actions: string[]) => Grant;
   // /**
   //  *  Adds an IAM policy statement for Mutation access to this GraphQLApi to an IAM principal's policy.
   //  *
@@ -158,13 +160,4 @@ export interface GraphQLAPIProvider extends IConstruct {
   //  * @param fields The fields to grant access to that are Subscriptions (leave blank for all).
   //  */
   grantSubscription: (grantee: IGrantable, ...fields: string[]) => Grant;
-}
-
-export interface APIIAMResourceProvider {
-  /**
-   * Return the Resource ARN
-   *
-   * @param api The GraphQL API to give permissions
-   */
-  resourceArns: (api: GraphQLAPIProvider) => string[];
 }