+++ packages/amplify-data-construct/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.10.1](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/data-construct@1.10.0...@aws-amplify/data-construct@1.10.1) (2024-09-16)
+
+**Note:** Version bump only for package @aws-amplify/data-construct
+
+# [1.10.0](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/data-construct@1.9.8...@aws-amplify/data-construct@1.10.0) (2024-09-06)
+
+### Features
+
+- **conversation:** add conversation transformer ([#2827](https://github.com/aws-amplify/amplify-category-api/issues/2827)) ([cee6aef](https://github.com/aws-amplify/amplify-category-api/commit/cee6aef1358293fe51909a64d1cf9941afc46aba))
+- **generation-transformer:** add generation transformer ([#2820](https://github.com/aws-amplify/amplify-category-api/issues/2820)) ([a86db4e](https://github.com/aws-amplify/amplify-category-api/commit/a86db4e40962565fb55b7262a7d771f21bacef2a))
+
+## [1.9.8](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/data-construct@1.9.7...@aws-amplify/data-construct@1.9.8) (2024-09-03)
+
+**Note:** Version bump only for package @aws-amplify/data-construct
+
+## [1.9.7](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/data-construct@1.9.6...@aws-amplify/data-construct@1.9.7) (2024-08-28)
+
+**Note:** Version bump only for package @aws-amplify/data-construct
+
+## [1.9.6](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/data-construct@1.9.5...@aws-amplify/data-construct@1.9.6) (2024-08-20)
+
+**Note:** Version bump only for package @aws-amplify/data-construct
+
+## [1.9.5](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/data-construct@1.9.4...@aws-amplify/data-construct@1.9.5) (2024-08-12)
+
+**Note:** Version bump only for package @aws-amplify/data-construct
+
 ## [1.9.4](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/data-construct@1.9.3...@aws-amplify/data-construct@1.9.4) (2024-08-01)
 
 **Note:** Version bump only for package @aws-amplify/data-construct